Experiment Utilities

Reproduce.@generate_config_funcs — Macro

@generate_config_funcs default_config

Generate a documented function default_config() which returns a default configuration Dict for an experiment. The default configuration Dict is built using the default_config argument, which should have the following form:

.
.
.
info"""
DOCUMENTATION
"""
DICTIONARY ELEMENTS
.
.
.

Where 'DOCUMENTATION' is a documentation for each element included in DICTIONARY ELEMENTS. 'DICTIONARY ELEMENTS' is a newline separated list of key => value pairs to be included in the default configuration dictionary. See the Examples section for more detail.

Examples

julia> @generate_config_funcs begin
    info"""
    Experiment details.
    --------------------
    - `seed::Int`: seed of RNG
    - `steps::Int`: Number of steps taken in the experiment
    """
    seed => 1
    steps => 200000

    info"""
    Agent details
    -------------
    - `latent_size::Int`: The size of the hidden layers in the RNN.
    """
    latent_size => 64

    info"""
    ### Optimizer details
    Flux optimizers are used. See flux documentation.
    - Parameters defined by the optimizer.
    """
    eta => 0.001

    info"""
    ### Learning update and replay details including:
    - Replay:
        - `replay_size::Int`: How many transitions are stored in the replay.
        - `warm_up::Int`: How many steps for warm-up (i.e. before learning begins).
    """
    replay_size => 10000
    warm_up => 1000

    info"""
    - Update details:
        - `lupdate::String`: Learning update name
        - `gamma::Float`: the discount for learning update.
        - `batch_size::Int`: size of batch
        - `truncation::Int`: Length of sequences used for training.
        - `update_wait::Int`: Time between updates (counted in agent interactions)
        - `target_update_wait::Int`: Time between target network updates (counted in agent interactions)
        - `hs_strategy::String`: Strategy for dealing w/ hidden state in buffer.
    """
    update => "QLearningMSE"
    gamma => 1.0
    batch_size=>32
    hist => 1
    epsilon => 0.1
    update_freq => 1
    target_update_wait => 100
end

julia> default_config()
Dict{String, Any} with 13 entries:
  "steps"              => 200000
  "warm_up"            => 1000
  "batch_size"         => 32
  "replay_size"        => 10000
  "eta"                => 0.001
  "hist"               => 1
  "target_update_wait" => 100
  "latent_size"        => 64
  "update"             => "QLearningMSE"
  "update_freq"        => 1
  "epsilon"            => 0.1
  "gamma"              => 1.0
  "seed"               => 1

Reproduce.@generate_working_function — Macro

@generate_working_function

Generate a documented function working_experiment() which wraps the main experiment function (main_experiment()) of a module and sets the arguments progress=true and testing=true, and uses the default experiment configuration (see @generate_config_funcs).

Reproduce.@param_from — Macro

@param_from param config_dict

Set the value of variable param to config_dict[string(param)]. There is also the capability to assign a type (or abstract type) you expect to recieve from the config for the key.

Examples

julia> d = Dict(
           "key1" => 1,
           "key2" => 2
       )
Dict{String, Int64} with 2 entries:
  "key2" => 2
  "key1" => 1

julia> @param_from key1 d
1

julia> @param_from key2::Int d
2

julia> println(key1, " ", key2)
1 2

julia> println(key1 + key2)
3