Experiment.jl

Reproduce.Experiment — Type

Experiment(dir, file, module_name, func_name, args_iter, config=nothing; modify_save_path=true)

This is a struct which encapsulates the idea of an experiment.

Arguments

dir: The directory associated w/ this experiment

file: The path to the file containing the experiment module

module_name: The module encapsulating your experiment (either a string or symbol)

func_name: The function inside the module which runs your experiment

args_iter: The settings your want to iterate over. This can be any iterator which returns (job_id, params)

config: [Optional] The path of the config file your experiment was built from. See Experiment(config_path, save_path="")

modify_save_path: If true and you are using ArgIterator or ArgLooper the argiter added to the experiment will have a new staticarg "save_dir" added.

Reproduce.Experiment — Type

Experiment(config_path, save_path="")

Build and experiment from either a toml file or a json.

Reproduce.add_experiment — Function

add_experiment(exp_dir, experiment_file, exp_module_name, exp_func_name, args_iter, hash, config)
add_experiment(exp::Experiment)

Set up details of experiment in exp_dir. This includes saving settings and config files, etc... It is recommended to use pre_experiment(exp::Experiment; kwargs...).

Reproduce.append_experiment_notes_file — Method

append_experiment_notes_file(exp_dir, experiment_file, exp_module_name, exp_func_name, settings_file, config_file, args_iter, config)

Write into an org file which contains the notes for the run experiments

Reproduce.create_experiment_dir — Method

create_experiment_dir(exp_dir; org_file=false, replace=false, tldr="")
create_experiment_dir(exp::Experiment; kwargs...)

This creates an experiment directory and sets up a notes file (if orgfile is true). It is recommended to use [`preexperiment(exp::Experiment; kwargs...)`](@ref).

Arguments

org_file=false: Indicator on whether to create an org file w/ notes on the job (experimental).

replace=false: If true the directory will be deleted. This is not recommended when calling createexperimentdir from multiple jobs.

tldr="": A TLDR which will be put into the org file.

Reproduce.pre_experiment — Method

pre_experiment(exp::Experiment; kwargs...)

Experiment setup phase. This helps deal with all the setup that needs to occur to setup an experiment folder.