parallel.jl

Reproduce.job — Method

job(experiment::Experiment; kwargs...)
job(experiment_file, exp_dir, args_iter; kwargs...)
job(experiment::Experiment, job_id; kwargs...)
job(experiment_file, exp_dir, args_iter, job_id; kwargs...)

Run a job specified by the experiment.

Reproduce.save_exception — Method

save_exception

This function saves an exception file with args:

config The job config that failed.

exc_file the file where the job should be saved.

job_id the id of the job being run (typically the idx of the job in the iterator).

exception the exception thrown by the job.

trace the stack trace of the raised exception.

Reproduce.get_comp_env — Method

get_comp_env

This derives the computational environment from the ENV variables. If in a slurm job the get_slurm_comp_env is used, if not get_local_comp_env is used.

Reproduce.get_local_comp_env — Method

get_local_comp_env

This checks to see if RP_TASK_ID is set in the environment. If so, a Task Job with ID=parse(Int, "RP_TASK_ID") will be returned. Otherwise, LocalParallel will be used. The kwargs (num_workers and threads_per_worker) give the job defaults for the number of parallel jobs and the number of threads per task. You can also use the ENV variables "RPNTASKS" and "RPCPUSPERTASK" to override these. The ENV variables will take precedence.

Reproduce.get_slurm_comp_env — Method

get_slurm_comp_env

This is significantly more complex than a local environment to enable using slurm task arrays efficiently.

In a job scheduled as:

sbatch -J test_rep_argiter --ntasks 4 --cpus-per-task 1 --mem-per-cpu=2000M --time=0:10:00 toml_parallel.jl configs/arg_iter_config.toml --path /home/mkschleg/scratch/reproduce

Then the comp_env will return a SlurmParallel env which uses srun to create julia instances.

For

sbatch -J test_rep_argiter -N 1 --ntasks 1 --cpus-per-task 4 --mem-per-cpu=2000M --time=0:10:00 toml_parallel.jl configs/arg_iter_config.toml --path /home/mkschleg/scratch/reproduce

The comp env will be a local parallel job using just default julia parallel utilities. It is necessary to make sure all resources are on single node if using this.

For

sbatch -J test_rep_argiter --array=1-4 --ntasks 4 --cpus-per-task 1 --mem-per-cpu=2000M --time=0:05:00 toml_parallel.jl configs/arg_iter_config.toml --path /home/mkschleg/scratch/reproduce

The comp env will be a SlurmTaskArray which will choose either LocalParallel, or SlurmParallel following the above protocol.

Notes:

If no SLURM_CPUS_PER_TASK is set then we assume a single cpu per task.

If you are re-running only parts of an array task you need to use "RPCUSTOMARRAYTASKCOUNT" to let Reproduce know what the original task array looked like to schedule the jobs correctly.

Reproduce.parallel_job_inner — Method

parallel_job_inner

Run a parallel job over the arguments presented by argsiter. `argsiter` can be a enumeration OR ArgIterator. Each job will be dedicated to a specific task. The experiment must save its own data! As this is not handled by this function (although could be added in the future.)