parse.jl

Reproduce.get_arg_iter — Method

get_arg_iter(cdict)
get_arg_iter(::Val{T}, cdict) where T

getargiter parses cdict to get the correct argument iterator. "argitertype" needs to have a string value in the cdict (where the cdict is the config dict, often from a config file). Reproduce has two iterators:

T=:iter: ArgIterator which does a grid search over arguments

T=:looper: ArgLooper which loops over a vector of dictionaries which can be loaded from an arg_file.

T=:iterV2: ArgIteratorV2 which is the second version of the original ArgIterator, and currently recommended.

To implement a custom argiter you must implement `Reproduce.getargiter(::Val{:symbol}, cdict)` where :symbol is the value argiter_type will take.

Reproduce.get_arg_iter — Method

get_arg_iter(::Val{:iterV2}, dict)

This is the function which parses ArgIteratorV2 from a config file dictionary. It expects the following nested dictionaries:

config: This has all the various components to help detail the expeirment (see parse_experiment_from_config for more details.)

arg_list_order::Vector{String}: inside the config dict is the order on which to do your sweeps. For example, if seed is first, the scheduler will make sure to run all the seeds for a particular setting before moving to the next set of parameters.

sweep_args: These are all the arguments that the args iter will sweep over (doing a cross product to produce all the parameters). See ArgIteratorV2 for supported features.

static_args: This is an optional component which contains all the arguments which are static. If not included, all elements in the top level of the dictionary will be assumed to be static args (excluding config and sweep_args).

Reproduce.get_save_backend — Method

get_save_backend(cdict)

Get the save_backend.

Reproduce.parse_experiment_from_config — Function

parse_experiment_from_config

This function creates an experiment from a config file.

args

config_path::String the path to the config.

[save_path::String] a save path which dictates where the base savedir for the job will be (prepend dict["config"]["save_dir"]).

kwargs

comp_env a computational environment which dispatchers when job is called.

The config file needs to be formated in a certain way. I use toml examples below:

[config]
save_dir="location/to/save" # will be prepended by save_path
exp_file="file/containing/experiment.jl" # The file containing your experiment function
exp_module_name = "ExperimentModule" # The module of your experiment in the experiment file
exp_func_name = "main_experiment" # The function to call in the experiment module.
arg_iter_type = "iterV2"

# These are specific to what arg_iter_type you are using 
[static_args]
...
[sweep_args]
...