Save

Reproduce.SaveManager — Type

DataManger

Abstract type for various filetype managers.

Reproduce.create_param_table — Method

create_param_table(dbm::DBManager, example_params)

Create the parameter table for an experiment.

Reproduce.create_results_tables — Method

create_results_table(dbm::DBManager, results)

Create the tables to store the results.

Reproduce.get_sql_name — Method

get_sql_name(name, param)

Returns the name used for a column. For a single value (i.e. String, Float, Integer, etc...) return name For a Tuple or a Vector return (name1, name2, ..., namen) For a NamedTuple return (name(prop1), name_(prop2), ...) where the props are sorted.

Reproduce.get_sql_schema — Method

get_sql_schema(name, param)

Reproduce.get_sql_type — Method

get_sql_type(x)

Return the corresponding SQL Type. This is to be used for params.

Reproduce.get_sql_value — Method

get_sql_value(x)

Return the value for the collection. If xi is a named tuple this returns the elements in alphabetical order wrt the element names. Otherwise, just return x.