Experiment.jl

Reproduce.Experiment — Type

Experiment

The structure used to embody a reproduce experiment. This is usually constructed through the parse_experiment_from_config, but can be used without config files.

dir: the base directory of the experiment (where the info files are saved).

file: The file containing the experiment function described by func_name and module_name

module_name: Module name containing the experiment function.

func_name: Function name of the experiment.

save_type: The save structure to deal with saving data passed by the experiment.

args_iter: The args iterator which contains the configs to pass to the experiment.

[confg]: The config file parsed to create the experiment (optional)

kwarg

[comp_env]: The computational environment used by the experiment.

Reproduce.experiment_dir_setup — Method

experiment_dir_setup

Sets up all the needed directories to run a parallel experiment.

Reproduce.experiment_save_init — Method

experiment_save_init(save::FileSave, exp::Experiment; kwargs...)
experiment_save_init(save::SQLSave, exp::Experiment; kwargs...)

Setups the necessary compoenents to save data for the jobs. This is run by pre_experiment. The FileSave creates the data directory where all the data is stored for an experiment. The SQLSave ensures the databases and tables are created necessary to successfully run an experiment.

Reproduce.post_experiment — Method

post_experiment

This doesn't do anything.

Reproduce.pre_experiment — Method

pre_experiment(exp::Experiment; kwargs...)
pre_experiment(file_save::FileSave, exp; kwargs...)
pre_experiment(sql_save::SQLSave, exp; kwargs...)

This function does all the setup required to successfully run an experiment. It is dispatched on the save structure in the experiment.

This function:

Creates the base experiment directory.

Runs experiment_save_init to initialize the details for each save type.

runs experiment_dir_setup