parallel.jl

Reproduce.job — Method

job(experiment::Experiment; kwargs...)
job(experiment_file, exp_dir, args_iter; kwargs...)
job(experiment::Experiment, job_id; kwargs...)
job(experiment_file, exp_dir, args_iter, job_id; kwargs...)

Run a job specified by the experiment.

Reproduce.parallel_job — Method

parallel_job

Run a parallel job over the arguments presented by argsiter. `argsiter` can be a enumeration OR ArgIterator. Each job will be dedicated to a specific task. The experiment must save its own data! As this is not handled by this function (although could be added in the future.)